Writing to the blockchain is fully working and accessible thorugh blockWriterService
