+++ lib/util/serializables/UserHasAccess.dart
@@ -4,13 +4,12 @@ import 'package:medrecs/util/serializables/iMedicalData.dart';
 
 class UserHasAccess extends iMedicalData {
   @override
-  final String entryType;
+  String entryType = "UserHasAccess";
   final int userID;
   final int userGrantedAccessID;
   final String date;
 
   UserHasAccess({
-    required this.entryType,
     required this.userID,
     required this.userGrantedAccessID,
     required this.date,
@@ -18,7 +17,6 @@ class UserHasAccess extends iMedicalData {
 
   factory UserHasAccess.fromJson(Map<String, dynamic> json, String type) {
     return UserHasAccess(
-      entryType: type,
       userID: json['userID'] as int,
       userGrantedAccessID: json['userGrantedAccessID'] as int,
       date: json['date'] as String,
@@ -70,4 +68,11 @@ class UserHasAccess extends iMedicalData {
   String getType() {
     return entryType;
   }
+
+  @override
+  Map<String, dynamic> toJson() => {
+        'userID': userID,
+        'userGrantedAccessID': userGrantedAccessID,
+        'date': date,
+      };
 }
